+++ tests/integration/pipelex/pipes/controller/pipe_batch/test_pipe_batch_simple.py
@@ -39,12 +39,12 @@ class TestPipeBatchSimple:
         concept_1 = ConceptFactory.make_from_blueprint(
             concept_code="TestConcept1",
             domain=domain,
-            blueprint=ConceptBlueprint(description="Lorem Ipsum"),
+            blueprint_or_string_description=ConceptBlueprint(description="Lorem Ipsum"),
         )
         concept_2 = ConceptFactory.make_from_blueprint(
             concept_code="TestConcept2",
             domain=domain,
-            blueprint=ConceptBlueprint(description="Lorem Ipsum"),
+            blueprint_or_string_description=ConceptBlueprint(description="Lorem Ipsum"),
         )
         concept_library = get_concept_library()
         concept_library.add_concepts([concept_1, concept_2])